refactor edge metamanager: introduce processors to distinguish messages
