+++ edge/pkg/common/modules/names.go
@@ -13,4 +13,6 @@ const (
 	DeviceTwinModuleName = "twin"
 	// EdgeHubModuleName name
 	EdgeHubModuleName = "websocket"
+	// MetaManagerModuleName name
+	MetaManagerModuleName = "metaManager"
 )
