+++ edge/pkg/metamanager/client/persistentvolume.go
@@ -10,7 +10,6 @@ import (
 	"github.com/kubeedge/beehive/pkg/core/model"
 	"github.com/kubeedge/kubeedge/edge/pkg/common/message"
 	"github.com/kubeedge/kubeedge/edge/pkg/common/modules"
-	"github.com/kubeedge/kubeedge/edge/pkg/metamanager"
 )
 
 // PersistentVolumesGetter is interface to get client PersistentVolumes
@@ -63,7 +62,7 @@ func (c *persistentvolumes) Get(name string, options metav1.GetOptions) (*api.Pe
 		return nil, fmt.Errorf("parse message to persistentvolume failed, err: %v", err)
 	}
 
-	if msg.GetOperation() == model.ResponseOperation && msg.GetSource() == metamanager.MetaManagerModuleName {
+	if msg.GetOperation() == model.ResponseOperation && msg.GetSource() == modules.MetaManagerModuleName {
 		return handlePersistentVolumeFromMetaDB(content)
 	}
 	return handlePersistentVolumeFromMetaManager(content)
