+++ edge/pkg/metamanager/client/secret.go
@@ -9,7 +9,6 @@ import (
 	"github.com/kubeedge/beehive/pkg/core/model"
 	"github.com/kubeedge/kubeedge/edge/pkg/common/message"
 	"github.com/kubeedge/kubeedge/edge/pkg/common/modules"
-	"github.com/kubeedge/kubeedge/edge/pkg/metamanager"
 )
 
 //SecretsGetter is interface to get client secrets
@@ -63,7 +62,7 @@ func (c *secrets) Get(name string) (*api.Secret, error) {
 	}
 
 	//op := msg.GetOperation()
-	if msg.GetOperation() == model.ResponseOperation && msg.GetSource() == metamanager.MetaManagerModuleName {
+	if msg.GetOperation() == model.ResponseOperation && msg.GetSource() == modules.MetaManagerModuleName {
 		return handleSecretFromMetaDB(content)
 	}
 	//else
